backend/requirements.txt
flask
flask-login
flask-sqlalchemy
werkzeug
sqlalchemy
pyjwt
